INGK-760 Initial implementation
