+++ ingenialink/utils/_utils.py
@@ -18,7 +18,7 @@ logger = ingenialogger.get_logger(__name__)
 
 POLLING_MAX_TRIES = 5  # Seconds
 
-__dtype_value: Dict[REG_DTYPE, Tuple[int, bool]] = {
+dtype_value: Dict[REG_DTYPE, Tuple[int, bool]] = {
     REG_DTYPE.U8: (1, False),
     REG_DTYPE.S8: (1, True),
     REG_DTYPE.U16: (2, False),
@@ -206,8 +206,8 @@ def convert_bytes_to_dtype(data: bytes, dtype: REG_DTYPE) -> Union[float, int, s
         ILValueError: If data can't be decoded in utf-8
     """
     signed = False
-    if dtype in __dtype_value:
-        bytes_length, signed = __dtype_value[dtype]
+    if dtype in dtype_value:
+        bytes_length, signed = dtype_value[dtype]
         data = data[:bytes_length]
 
     if dtype == REG_DTYPE.FLOAT:
@@ -246,6 +246,6 @@ def convert_dtype_to_bytes(data: Union[int, float, str, bytes], dtype: REG_DTYPE
         return data.encode("utf_8")
     if not isinstance(data, int):
         raise ValueError(f"Expected data of type int, but {type(data)}")
-    bytes_length, signed = __dtype_value[dtype]
+    bytes_length, signed = dtype_value[dtype]
     data_bytes = data.to_bytes(bytes_length, byteorder="little", signed=signed)
     return data_bytes
